api/requirements.txt
fastapi>=0.115
email-validator>=2.0
uvicorn[standard]>=0.30
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.20
asyncpg>=0.29
python-jose[cryptography]>=3.3
bcrypt>=4.0
authlib>=1.3
httpx>=0.27
pydantic-settings>=2.0
alembic>=1.13
python-multipart>=0.0.9
aiofiles>=23.0
respx>=0.21

# Test deps
pytest>=8
pytest-asyncio>=0.23
